toro/requirements.txt
discord.py
urllib3